Expose public #open method
There is currently no way to get a persistent file handle from the
public interface. You sometimes want to do this, e.g. for defining a log
device.

When a block is not passed to File.open, it returns the open file
instance. MemoryFileSystem was not doing this, making the interfaces
inconsistent.